+++ Lib/fontbakery/profiles/adobefonts.py
@@ -56,6 +56,7 @@ PROFILE = {
         "interpolation_issues",
         "legacy_accents",
         "math_signs_width",
+        "missing_small_caps_glyphs",
         "no_debugging_tables",
         "no_mac_entries",
         "render_own_name",